refactor(core): 将 PackageManager 从 __main__ 迁移至 utils 模块并优化相关导入
